#3 Improve documentation
Explain how to load with a Configuration. I did not find an explanation for this in "Pharo by Example" or "Deep into Pharo", so we should explain it here. MorphiDraw has I see it the intention to help beginners with Pharo.
Replace also workspace with Playground as Newcomers to Pharo will be more convenient with "Playground"